Add working version of fullstack app

+++ controllers/auth.controller.js
@@ -4,26 +4,26 @@ const getImageFileType = require('../utils/getImageFileType');
 const fs = require('fs');
 
 exports.register = async (req, res) => {
-try {
-  const { login, password} = req.body;
-  const fileType = req.file ? await getImageFileType(req.file) : 'unknown';
-  if(login && typeof login === 'string' && password && typeof password === 'string' && req.file && ['image/png', 'image/jpeg', 'image/gif'].includes(fileType) ) {
-    const userWithlogin = await User.findOne({ login });
-    if (userWithlogin) {
-       return res.status(409).send({ message: 'User with this login already exists'});
-    }
-
-    const user = await User.create({ login, password: await bcrypt.hash(password, 10), avatar: req.file.filename });
-    res.status(201).send({ message: 'User created ' + user.login})
-    } else {
-        fs.unlinkSync(req.file.path);
-        res.status(400).send({ message: 'Bad request'});
-    }
+    try {
+      const { login, password, number} = req.body;
+      const fileType = req.file ? await getImageFileType(req.file) : 'unknown';
+      if(login && typeof login === 'string' && password && typeof password === 'string' && req.file && ['image/png', 'image/jpeg', 'image/gif'].includes(fileType) && number && typeof parseInt(number) === 'number') {
+        const userWithlogin = await User.findOne({ login });
+        if (userWithlogin) {
+           return res.status(409).send({ message: 'User with this login already exists'});
+        }
     
-    } catch (err) {
-        res.status(500).send({message: err.message});
+        const user = await User.create({ login, password: await bcrypt.hash(password, 10), avatar: req.file.filename, number });
+        res.status(201).send({ message: 'User created ' + user.login})
+        } else {
+            fs.unlinkSync(req.file.path);
+            res.status(400).send({ message: 'Bad request'});
+        }
+        
+        } catch (err) {
+            res.status(500).send({message: err.message});
+        }
     }
-}
 
 exports.login = async (req,res) => {
     try {
@@ -35,6 +35,7 @@ exports.login = async (req,res) => {
             } else {
                 if(bcrypt.compareSync(password, user.password)) {
                     req.session.user = {login: user.login, id: user._id}
+                    req.session.authorized = true;
                     res.status(200).send({message: 'Login successful'});
                 } else {
                     res.status(400).send({message: 'Login or password are inncorect'});
@@ -49,7 +50,8 @@ exports.login = async (req,res) => {
 }
 
 exports.getUser = async (req, res) => {
-    res.send({ message: 'I\'m logged!' })
+    
+    res.json({ loggedIn: true, user: req.session.user });
 }
 
 exports.logout = async (req, res) => {
